requirements.txt
flask
deepgram-sdk
python-dotenv
openai
